Refactor popular.dart and popular_item.dart files
